Don't delete old CuckooLock objects
On the one hand, we can't simply delete these since other threads might
still hold a reference and try to access these afterwards. On the other,
shared_ptr is not an option because it turns into the bottleneck.

As a first step, simply don't delete old locks and live with the memory
leak.